# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API documentation for managing Akun resources",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/profil": {
            "get": {
                "tags": [
                    "Profil"
                ],
                "summary": "Get list of Profil",
                "description": "Display a listing of Profil.",
                "operationId": "443cfc4a0263ea00b547094ea694b2de",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Profil"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Profil"
                ],
                "summary": "Create a new Profil",
                "description": "Store a new Profil.",
                "operationId": "4deb5a5dba5bf59b3d4965fe303be526",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "username": {
                                        "type": "integer",
                                        "example": "jhondoe"
                                    },
                                    "password": {
                                        "description": "User password, minimum 8 characters",
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "id_role": {
                                        "type": "string",
                                        "example": "role_123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Profil created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Profil"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/profil/{id}": {
            "get": {
                "tags": [
                    "Profil"
                ],
                "summary": "Get a Profil by ID",
                "description": "Display a Profil by ID.",
                "operationId": "274e688502ca391d35a011a30c94643d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Profil"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Profil not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Profil"
                ],
                "summary": "Update a Profil by ID",
                "description": "Update an existing Profil.",
                "operationId": "14064df4f4eb7a597bfeac43b53603e4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama": {
                                        "type": "string",
                                        "example": "Jane Doe"
                                    },
                                    "username": {
                                        "type": "integer",
                                        "example": "janedoe"
                                    },
                                    "password": {
                                        "description": "User password, minimum 8 characters",
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "id_role": {
                                        "type": "string",
                                        "example": "role_456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profil updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Profil"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Profil not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Profil"
                ],
                "summary": "Delete a Profil by ID",
                "description": "Delete a Profil.",
                "operationId": "ac75740d469e8db025344ffbf045ef03",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Profil deleted"
                    },
                    "404": {
                        "description": "Profil not found"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "User login",
                "description": "Authenticate user and return token",
                "operationId": "a677d3288ec8adfa5dfb15b56f8079ab",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "password"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "user123"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Login successful"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "1|abc123..."
                                        },
                                        "data": {
                                            "properties": {
                                                "id_profil": {
                                                    "type": "string",
                                                    "example": "id123"
                                                },
                                                "username": {
                                                    "type": "string",
                                                    "example": "user123"
                                                },
                                                "nama": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "id_role": {
                                                    "type": "string",
                                                    "example": "role123"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Login failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Login failed, invalid credentials"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "User logout",
                "description": "Revoke user's token",
                "operationId": "2216f76882e0e7725c4d58d9a7930ab9",
                "responses": {
                    "200": {
                        "description": "Logged out successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Logged out successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Profil": {
                "properties": {
                    "id_profil": {
                        "type": "string",
                        "example": "profil_123"
                    },
                    "id_role": {
                        "type": "string",
                        "example": "role_123"
                    },
                    "kode": {
                        "type": "integer",
                        "example": 101
                    },
                    "nama": {
                        "type": "string",
                        "example": "John Doe"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}